Better error handling for the conf header
Adds several rules in the lexxer that emit much better errors for the
conf header. When rules are broken for either part of the conf header,
instead of blowing up and jumping down to a general error, each section
now has its own error state. More descriptive output could still be
added but this is a good start.