Goodbye Doxygen.
For now, no documentation generator. Eventually, most of this code will
be pulled into a library to abstract the CLI functionality from the
actual parser and lexxer. When that happens, docs will likely be
generated by Docorium, not Doxygen. I mean, if Docorium is good enough
for libgit2, it sure as hell is good enough for libmt. ^_^