Adds invalid string errors
Beefs up the mt_error handler with invalid string errors. These all take
place in the construction of the note (the only place in the parser
where strings are set). As notes are immutable, with respect to string
and fret numbers, the error checking only needs to take place in the
constructors.