Refactors Makefile
This is adding some basic knowlege of how Make works to the Makefile.
The previous Makefile would actually fail to make almost every time. I
tested it on a special case where everything looked like it was working
right. Six hours alone with the GNU Make guide later and things are
different.