Fixes another make inconsistency
Running `make` would succeed the first time but fail for all subsequent
runs. This was due to the way that the manifest of marktab srcs was
generated. The makefile was using a simple wildcard to grab all the *.c
files in the src/ directory. A problem arose once make had already been
run. The flex and bison steps added two *.c files to the src/ directory.
This caused the wildcard to pick those up the second time through and,
subsequently, add them to the final gcc line twice.